+++ app/src/main/java/com/example/login_ui/util/UserBean.java
@@ -1,18 +1,18 @@
 package com.example.login_ui.util;
 
 public class UserBean {
-    private Long id;          //用户ID，不显示
+    private int id;          //用户ID，不显示
     private String nickName;     //昵称，显示用
 
     public void setNickName(String nickName) {
         this.nickName = nickName;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
